refactor: split provider listing out of loader
Move list_providers() into a new discovery.py so loader.py holds only what a
backend needs to resolve dynamic metadata; the unknown-provider error now
enumerates entry-point names inline instead of via list_providers. Update the
CLI, tests, and docs to the new location.
